Initial Setup / Swipe Detection
I set up folders for my setup so I won't clash with others. I also created a swipe detection that uses Raycast to hit whatever objects and detects if that object is theirs or not. There are also directional use cases as it needs the correct direction in order to activate.