Priority system & puzzle tag
Created a static variable in the script GameEventManager. It is a script that talks to everything and going forward, anything related to sharing with every scene can be put into here. For now it is just a single bool that just says if or if not the touch, is touching a puzzle.

I also created a tag that sets if the thing it is touching has a puzzle tag or not. Because raycasts detects every object, so I had to assign tags. Just remember to set puzzles with the tag of puzzles