Fixed flipping / initial testing of page 3
I fixed the weird bug that keeps flickering to the left and right at the start. Now it works perfectly fine. Now it's time for more pages which means to test out if my code truly works. (Spoilers it doesn't) Time for fixing the damn stupid code for page 3 so I don't hard code myself to death